Migrate solver engine from Timefold 1.33.0 to Timefold 2.2.0
Bump timefold-solver-bom to 2.2.0 and adapt to the 2.0 API removals:

- @PlanningVariable(nullable=true) -> allowsUnassigned=true
- HardSoftScore package flattened (...api.score.buildin.hardsoft -> ...api.score)
- scores are long-based: hardScore()/softScore() return long (AspApp display vars)
- @PlanningId -> ...api.domain.common; ConstraintVerifier -> ...core.api.score.stream.test
- timefold-solver-test artifact removed (folded into core); drop the dependency
- ConstraintCollectors min/max(fn, Comparator) overload removed -> single-arg form
- ForbiddenCellSelectionFilter: impl ChangeMove -> SelectorBasedChangeMove<Solution>,
  ScoreDirector moved to impl package
- JsonSolutionFileIO: timefold-solver-jackson now uses Jackson 3 (tools.jackson.*,
  immutable ObjectMapper) -> build via JsonMapper.builder()...build()
- solver/benchmark XML: drop removed constraintStreamImplType (BAVET is the only
  engine); environmentMode REPRODUCIBLE -> NO_ASSERT

Add the always-on structural constraint `partialAssignment`: 2.0's Local Search
nulls a single planning variable to cheaply shed a penalty, stranding shifts in
partial (one-variable-set) states that are invisible to every per-shift constraint
and form local-optima traps, leaving the solve ~3 hard short of feasibility.
Penalizing the partial state restores the gradient out of those traps and the
production solve reaches 0hard again. It is 0 on any fully-assigned or
fully-unassigned schedule, so the legacy oracle, characterization golden-masters,
and exact-match differential cases are unchanged and feasibility-equivalence holds.

Tests: 59/59 pass, including the 4-min solve-and-grade feasibility gate.

+++ src/main/java/org/betaiotazeta/autoshiftplanner/AspApp.java
@@ -912,8 +912,8 @@ public class AspApp extends javax.swing.JFrame {
                 Solution interimSolution = chunks.get(chunks.size() - 1);
                 
                 // Display the result
-                int interimHardScore = interimSolution.getScore().hardScore();
-                int interimSoftScore = interimSolution.getScore().softScore();
+                long interimHardScore = interimSolution.getScore().hardScore();
+                long interimSoftScore = interimSolution.getScore().softScore();
                 String text = (interimHardScore + " hard, " + interimSoftScore + " soft");
                 score_jLabel.setText(text);
 
@@ -960,8 +960,8 @@ public class AspApp extends javax.swing.JFrame {
                     solution = solvedSolution;
 
                     // Display the result
-                    int solvedHardScore = solvedSolution.getScore().hardScore();
-                    int solvedSoftScore = solvedSolution.getScore().softScore();
+                    long solvedHardScore = solvedSolution.getScore().hardScore();
+                    long solvedSoftScore = solvedSolution.getScore().softScore();
                     String message = "Solved score: " + solvedHardScore + " hard, " + solvedSoftScore + " soft.";
                     JOptionPane.showMessageDialog(aspApp, message, "Information", JOptionPane.INFORMATION_MESSAGE);
